Simplify packages
The desk package was just a pass through so
really not required, the function APIs were
the same signature as the store just with
different names, once the names were aligned
there was no point in keeping the separate
package.

At the same time (because I am a glutton for
punishment), the custom ID was switched to the
underlying integer type because there was no
real advantage to the separate type.
Similarly desk status has been removed as
the current customers do not have a need for
it. It would be simple to add back if needed
in the future.