Simplify configuration
A big problem with administering the system
(after users which is now solved), is
managing desks. With this change desks are
simply defined in a text file rather than
in the database.

If performance becomes an issue (really?)
in the future then initialisation of the
in-mem desk store (by reading the file)
could be done in parallel with the sqlite
database connection. It would be nice to
ever have that problem with this system.